Feat: 기타 수정
- 기존 createdDateTime 순으로 정렬했으나, Id 순으로 정렬하도록 변경함
- 심화 개인 과제 피드백 반영: entity 수정 시 더티 체킹 사용하는 쪽으로 일관성 있게 수정
- 자원 소유자 권한 확인 시 if가 아닌 Preconditions의 check 함수 사용
- Entity에서 DTO를 모르도록 확장함수 위치 변경
- 더 고민할 부분
  - QueryDSL과 Pageable 함께 사용한 ToDoCard 목록 조회 메서드에서 복잡한 쿼리가 나가고 있는 듯함
    - 쿼리 다시 확인하고 고민할 것
