+++ kwakjoonseon/kjs/week10/src/main/kotlin/kotlinassignment/week10/domain/comment/service/CommentService.kt
@@ -10,10 +10,9 @@ import org.springframework.data.domain.Pageable
 interface CommentService {
 
     /**
-     * TODO - 주석 정리
      * @param toDoCardId (type: Long) Comment와 연관된 ToDoCard의 id
-     * @param pageable (type: CommentCreateRequest)
-     * @return (type: Page<CommentResponse>)
+     * @param pageable (type: CommentCreateRequest) Comment 목록을 가져올 때 필요한 page number, page size, sort 등의 정보를 담은 pageable
+     * @return (type: Page<CommentResponse>) Comment 데이터 목록을 Page 형태로 담아 반환
      */
     fun getCommentList(toDoCardId: Long, pageable: Pageable): Page<CommentResponse>
