+++ kwakjoonseon/kjs/week10/src/main/kotlin/kotlinassignment/week10/domain/toDoCard/dto/ToDoCardResponse.kt
@@ -1,12 +1,24 @@
 package kotlinassignment.week10.domain.toDoCard.dto
 
+import kotlinassignment.week10.domain.toDoCard.model.ToDoCard
 import java.time.LocalDateTime
 
 data class ToDoCardResponse(
     val id: Long,
     val title: String,
     val description: String?,
-    val memberNickname: String, // TODO nickname 등으로 수정 필요 - entity까지 수정 필요해짐
+    val memberNickname: String,
     val createdDateTime: LocalDateTime,
     val isComplete: Boolean,
 )
+
+fun ToDoCard.toResponse(): ToDoCardResponse {
+    return ToDoCardResponse(
+        id = this.id!!,
+        title = this.title,
+        description = this.description,
+        memberNickname = this.member.nickname,
+        createdDateTime = this.createdDateTime,
+        isComplete = this.isComplete,
+    )
+}
