+++ kwakjoonseon/kjs/week10/src/main/kotlin/kotlinassignment/week10/domain/toDoCard/repository/CustomToDoCardRepository.kt
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Pageable
 
 interface CustomToDoCardRepository {
 
-    fun findAllFilteringByTitleOrUserNameWithSortOrder(
+    fun findAllFilteringByTitleOrUserName(
         title: String?,
         memberNickname: String?,
         pageable: Pageable,
